EES-5062 Add Filter/IndicatorSequence columns to ReleaseFile
