+++ src/GovUk.Education.ExploreEducationStatistics.Content.Model/ReleaseFile.cs
@@ -1,5 +1,6 @@
 #nullable enable
 using System;
+using System.Collections.Generic;
 
 namespace GovUk.Education.ExploreEducationStatistics.Content.Model
 {
@@ -20,5 +21,21 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Model
         public string? Summary { get; set; }
 
         public int Order { get; set; }
+
+        public List<FilterSequenceEntry>? FilterSequence { get; set; }
+
+        public List<IndicatorGroupSequenceEntry>? IndicatorSequence { get; set; }
+
     }
+
+    public abstract record SequenceEntry<TEntry, TChild>(TEntry Id, List<TChild> ChildSequence);
+
+    public record FilterSequenceEntry(Guid Id, List<FilterGroupSequenceEntry> ChildSequence) :
+        SequenceEntry<Guid, FilterGroupSequenceEntry>(Id, ChildSequence);
+
+    public record FilterGroupSequenceEntry(Guid Id, List<Guid> ChildSequence) :
+        SequenceEntry<Guid, Guid>(Id, ChildSequence);
+
+    public record IndicatorGroupSequenceEntry(Guid Id, List<Guid> ChildSequence) :
+        SequenceEntry<Guid, Guid>(Id, ChildSequence);
 }
