+++ src/GovUk.Education.ExploreEducationStatistics.Data.Services.Tests/Utils/FiltersMetaViewModelBuilderTests.cs
@@ -1,6 +1,7 @@
 ﻿#nullable enable
 using System;
 using System.Collections.Generic;
+using GovUk.Education.ExploreEducationStatistics.Content.Model;
 using GovUk.Education.ExploreEducationStatistics.Data.Model;
 using GovUk.Education.ExploreEducationStatistics.Data.ViewModels.Meta;
 using Xunit;
