+++ src/GovUk.Education.ExploreEducationStatistics.Data.Services.Tests/Utils/IndicatorsMetaViewModelBuilderTests.cs
@@ -2,6 +2,7 @@
 using System;
 using System.Collections.Generic;
 using GovUk.Education.ExploreEducationStatistics.Common.Model;
+using GovUk.Education.ExploreEducationStatistics.Content.Model;
 using GovUk.Education.ExploreEducationStatistics.Data.Model;
 using GovUk.Education.ExploreEducationStatistics.Data.ViewModels.Meta;
 using Xunit;
